requirements.txt
albucore

albumentations==2.0.8

annotated-types==0.7.0

anyio==4.11.0

appdirs==1.4.4

archspec

argcomplete==3.6.3

asgiref==3.10.0

attr==0.3.1

attrs

azure-core==1.36.0

azure-storage-blob==12.27.1

black==25.11.0

bleach==5.0.1

boltons

boto==2.49.0

boto3==1.40.70

botocore==1.40.70

Brotli

cachetools==6.2.1

certifi==2025.10.5

cffi==2.0.0

charset-normalizer==3.4.4

click==8.3.0

colorama==0.4.6

cryptography==46.0.3

datamodel-code-generator==0.26.1

defusedxml==0.7.1

distro

Django==5.1.14

django-annoying==0.10.6

django-cors-headers==4.7.0

django-csp==3.7

django-debug-toolbar==3.2.1

django-environ==0.10.0

django-extensions==3.2.3

django-filter==24.3

django-migration-linter==5.2.0

django-model-utils==4.1.1

django-ranged-fileresponse==0.1.2

django-rq==2.10.3

django-storages==1.12.3

django-user-agents==0.4.0

djangorestframework==3.15.2

djangorestframework_simplejwt==5.5.1

dnspython==2.8.0

drf-dynamic-fields==0.3.0

drf-flex-fields==0.9.5

drf-generators==0.3.0

drf-spectacular==0.28.0

email-validator==2.3.0

eval_type_backport

expiringdict==1.2.2

flask

Faker==37.12.0

filelock==3.20.0

frozendict

genson==1.3.0

google-api-core==2.28.1

google-auth==2.43.0

google-cloud-appengine-logging==1.7.0

google-cloud-audit-log==0.4.0

google-cloud-core==2.5.0

google-cloud-logging==3.12.1

google-cloud-storage==2.19.0

google-crc32c==1.7.1

google-resumable-media==2.7.2

googleapis-common-protos==1.72.0

grpc-google-iam-v1==0.14.3

grpcio==1.76.0

grpcio-status==1.76.0

h11==0.16.0

h2

hpack

httpcore==1.0.9

httpx==0.28.1

hyperframe

idna==3.11

ijson==3.4.0.post0

importlib_metadata==8.7.0

importlib_resources

inflect==5.6.2

inflection==0.5.1

iniconfig==2.3.0

isodate==0.7.2

isort==5.13.2

Jinja2==3.1.6

jiter==0.12.0

jmespath==1.0.1

joblib==1.5.2

jsf==0.11.2

jsonpatch

jsonpointer

jsonschema

jsonschema-specifications

label-studio==1.21.0

label-studio-sdk==2.0.4

launchdarkly-server-sdk==8.2.1
lockfile==0.12.2

lxml==6.0.2

lxml_html_clean==0.4.3

markdown-it-py==4.0.0

MarkupSafe==3.0.3

mdurl==0.1.2
mypy_extensions==1.1.0

nltk==3.9.2

numpy==1.26.4

openai==1.109.1

opencv-python==4.11.0.86

opencv-python-headless==4.11.0.86

onnxruntime==1.19.2

opentelemetry-api==1.38.0

ordered-set==4.0.2

packaging==25.0

pandas==2.3.3

pathspec==0.12.1

pillow==12.0.0

pkgutil_resolve_name

platformdirs==4.5.0

pluggy==1.6.0

proto-plus==1.26.1

protobuf==6.33.0

psycopg2-binary==2.9.10

pyarrow==18.1.0

pyasn1==0.6.1

pyasn1_modules==0.4.2

pyboxen==1.3.0

pycparser

pydantic==2.12.4

pydantic_core==2.41.5

Pygments==2.19.2

PyJWT==2.10.1

pyRFC3339==2.1.0

PySocks

pytest==8.4.2

python-dateutil==2.9.0.post0

python-json-logger==2.0.4

pytokens==0.3.0

pytz==2022.7.1

PyYAML==6.0.3

PyMuPDF==1.23.8

redis==5.2.1

referencing

regex==2025.11.3

requests==2.32.5

requests-file==3.0.1

requests-mock==1.12.1

rich==14.2.0

rpds-py

rq==1.16.2

rsa==4.9.1

rstr==3.2.2

ruamel.yaml

ruamel.yaml.clib

rules==3.4

s3transfer==0.14.0

scipy==1.16.3

semver==3.0.4

sentry-sdk==2.44.0

setuptools==80.9.0

simsimd==6.5.3

six==1.17.0

smart_open==7.5.0

sniffio==1.3.1

sqlparse==0.5.3

stringzilla==4.2.3

tldextract==5.3.0

toml==0.10.2

tqdm

torch==2.5.1

truststore

typing-inspection==0.4.2

typing_extensions==4.15.0

tzdata==2025.2

ua-parser==1.0.1

ua-parser-builtins==0.18.0.post1

ujson==5.11.0

uritemplate==4.2.0

urllib3==2.5.0

user-agents==2.2.0

ultralytics==8.3.228

webencodings==0.5.1

wheel==0.40.0

win_inet_pton

wrapt==2.0.1

xmljson==0.2.1

zipp

zstandard==0.23.0
